Move `PublishComment` command
